test: Add test case for number/float field type.

+++ plugins/wpe-content-model/includes/content-registration/custom-post-types-registration.php
@@ -207,10 +207,10 @@ function map_html_field_type_to_graphql_field_type( string $field_type ): ?strin
 		case 'textarea':
 		case 'string':
 		case 'date':
-		case 'media': // ?
+		case 'media':
 			return 'String';
 		case 'number':
-			return 'Int';
+			return 'Float';
 		default:
 			return null;
 	}
